gitserver: move repo_count metrics to worker
