+++ cmd/gitserver/server/servermetrics.go
@@ -1,7 +1,6 @@
 package server
 
 import (
-	"context"
 	"os/exec"
 	"syscall"
 	"time"
@@ -63,47 +62,6 @@ func (s *Server) RegisterMetrics(db dbutil.DB, observationContext *observation.C
 		var stat syscall.Statfs_t
 		_ = syscall.Statfs(s.ReposDir, &stat)
 		return float64(stat.Blocks * uint64(stat.Bsize))
-	})
-	prometheus.MustRegister(c)
-
-	c = prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "src_gitserver_repo_last_error_total",
-		Help: "Number of repositories whose last_error column is not empty.",
-	}, func() float64 {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var count int64
-		err := db.QueryRowContext(ctx, `
-			SELECT COUNT(*) FROM gitserver_repos AS g
-			INNER JOIN repo AS r ON g.repo_id = r.id
-			WHERE g.last_error IS NOT NULL AND r.deleted_at IS NULL
-		`).Scan(&count)
-		if err != nil {
-			s.Logger.Error("failed to count repository errors", log.Error(err))
-			return 0
-		}
-		return float64(count)
-	})
-	prometheus.MustRegister(c)
-
-	c = prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "src_gitserver_repo_count",
-		Help: "Number of repos.",
-	}, func() float64 {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var count int64
-		err := db.QueryRowContext(ctx, `
-			SELECT COUNT(*) FROM repo AS r
-			WHERE r.deleted_at IS NULL
-		`).Scan(&count)
-		if err != nil {
-			s.Logger.Error("failed to count repositories", log.Error(err))
-			return 0
-		}
-		return float64(count)
 	})
 	prometheus.MustRegister(c)
